use i32 instead of i64 everywhere

+++ src/bot.rs
@@ -30,7 +30,7 @@ pub struct BotConfig {
     pub tester_group: ChatId,
     pub reviewee_group: ChatId,
     pub genesis_group: ChatId,
-    pub serial: i64,
+    pub serial: i32,
 }
 
 /// Github Bot state
